Windows
+ Windowsbuilds für g++ 4.6 von MinGW (und zukünftig auch 4.7, wenn der
  mal mit Qt tut). EXE liegt ab sofort im Repo, weil Windowsnutzer zu doof
  zum Selbstkompilieren sind und so.

+++ texture_management.cpp
@@ -126,7 +126,7 @@ static GLuint load_bmp_texture(const char *path)
 
 
     unsigned char *content = new unsigned char[lof - ofs];
-    if ((long)fread(content, 1, lof - ofs, fp) < lof - ofs)
+    if ((long)fread(content, 1, lof - ofs, fp) < lof - (long)ofs)
     {
         delete content;
         fclose(fp);
